+++ src/models.ts
@@ -4,6 +4,7 @@ import RawResource from "./lib/resources/raw/raw-resource";
 import SimDataResource from "./lib/resources/simdata/simdata-resource";
 import StringTableResource from "./lib/resources/stbl/stbl-resource";
 import CombinedTuningResource from "./lib/resources/combined-tuning/combined-tuning-resource";
+import DdsImageResource from "./lib/resources/dds-image/dds-image-resource";
 
 export {
   Package,
@@ -11,5 +12,6 @@ export {
   XmlResource,
   SimDataResource,
   StringTableResource,
-  CombinedTuningResource
+  CombinedTuningResource,
+  DdsImageResource
 }