+++ src/types.ts
@@ -7,10 +7,12 @@ import type { WritableModelCreationOptions, WritableModelFromOptions } from "./l
 import type { RawResourceCreationOptions } from "./lib/resources/raw/raw-resource";
 import type { SimDataResourceCreationOptions } from "./lib/resources/simdata/simdata-resource";
 import type { XmlResourceFromOptions } from "./lib/resources/xml/xml-resource";
+import type { DdsImageResourceCreationOptions } from "./lib/resources/dds-image/dds-image-resource";
 
 export {
   // options
   BinaryFileReadingOptions,
+  DdsImageResourceCreationOptions,
   PackageFileReadingOptions,
   XmlExtractionOptions,
   WritableModelCreationOptions,